Restructured styles into its own directory
